requirements.txt
streamlit
langchain
langchain-groq
pydantic
pdfminer.six
python-dotenv